Generalized Raga to multimass models and added it to the AMUSE interface

+++ src/math_linalg.cpp
@@ -1065,7 +1065,7 @@ std::vector<Triplet> SparseMatrix<NumT>::values() const
     result.reserve(sp->nz);
     size_t col = 0;
     for(size_t k=0; k<sp->nz; k++) {
-        while(col<sp->size2 && sp->p[col+1]<=k)
+        while(col < sp->size2 && sp->p[col+1] <= static_cast<int>(k))
             col++;
         result.push_back(Triplet(sp->i[k], col, sp->data[k]));
     }
